third_party/WebKit/LayoutTests/external/wpt/cookies/samesite/form-get-blank-reload-expected.txt

This is a testharness.js-based test.
PASS Reloaded same-host top-level form GETs are strictly same-site
PASS Reloaded subdomain top-level form GETs are strictly same-site
FAIL Reloaded cross-site top-level form GETs are laxly same-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.842691446233597" but got (undefined) undefined
Harness: the test ran to completion.
